# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "excerptlab",
    "contact": {}
  },
  "host": "api.excerptlab.com",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admin/listUser": {
      "get": {
        "summary": "listUser",
        "tags": [
          "admin"
        ],
        "operationId": "listUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/updateUser/{userId}": {
      "put": {
        "summary": "update user by admin",
        "tags": [
          "admin"
        ],
        "operationId": "updateuserbyadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateuserbyadminrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/deleteUser/{userId}": {
      "delete": {
        "summary": "deleteUser",
        "tags": [
          "admin"
        ],
        "operationId": "deleteUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/createTopic": {
      "post": {
        "summary": "createTopic",
        "tags": [
          "admin"
        ],
        "operationId": "createTopic",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createTopicRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/listTopics": {
      "get": {
        "summary": "listTopics",
        "tags": [
          "admin"
        ],
        "operationId": "listTopics",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/updateTopic/{topicId}": {
      "put": {
        "summary": "updateTopic",
        "tags": [
          "admin"
        ],
        "operationId": "updateTopic",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateTopicRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/deleteTopic/{topicId}": {
      "delete": {
        "summary": "deleteTopic",
        "tags": [
          "admin"
        ],
        "operationId": "deleteTopic",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/createQuestion": {
      "post": {
        "summary": "createQuestion",
        "tags": [
          "admin"
        ],
        "operationId": "createQuestion",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createQuestionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/listQuestions": {
      "get": {
        "summary": "listQuestions",
        "tags": [
          "admin"
        ],
        "operationId": "listQuestions",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "topicId",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/updateQuestion/{questionId}": {
      "put": {
        "summary": "updateQuestion",
        "tags": [
          "admin"
        ],
        "operationId": "updateQuestion",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateQuestionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/deleteQuestion/{questionId}": {
      "delete": {
        "summary": "deleteQuestion",
        "tags": [
          "admin"
        ],
        "operationId": "deleteQuestion",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/getQuestionById/{questionId}": {
      "get": {
        "summary": "getQuestionById",
        "tags": [
          "admin"
        ],
        "operationId": "getQuestionById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/deleteFeedback/{feedbackId}": {
      "delete": {
        "summary": "delete Feedback",
        "tags": [
          "admin"
        ],
        "operationId": "deleteFeedback",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/createHint": {
      "post": {
        "summary": "create hint",
        "tags": [
          "admin"
        ],
        "operationId": "createhint",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createhintrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/updateHint/{hintId}": {
      "put": {
        "summary": "update hint",
        "tags": [
          "admin"
        ],
        "operationId": "updatehint",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "hintId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updatehintrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/deleteHint/{hintId}": {
      "delete": {
        "summary": "delete hint",
        "tags": [
          "admin"
        ],
        "operationId": "deletehint",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "hintId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/listHints": {
      "get": {
        "summary": "list hints",
        "tags": [
          "admin"
        ],
        "operationId": "listhints",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/uploadFile": {
      "post": {
        "summary": "upload files",
        "tags": [
          "admin"
        ],
        "operationId": "uploadfiles",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "files",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "updateuserbyadminrequest": {
      "title": "updateuserbyadminrequest",
      "example": {
        "name": "abcd",
        "email": "abcd@yopmail.com"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email"
      ]
    },
    "createTopicRequest": {
      "title": "createTopicRequest",
      "example": {
        "name": "trumpet",
        "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU",
        "level": 6,
        "rowNo": 6,
        "position": "right",
        "prompt": [
          {
            "type": "text",
            "value": "hi"
          },
          {
            "type": "image",
            "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
          }
        ]
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "rowNo": {
          "type": "integer",
          "format": "int32"
        },
        "position": {
          "type": "string"
        },
        "prompt": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Prompt"
          }
        }
      },
      "required": [
        "name",
        "image",
        "level",
        "rowNo",
        "position",
        "prompt"
      ]
    },
...
    "updateTopicRequest": {
      "title": "updateTopicRequest",
      "example": {
        "name": "rhythm-1",
        "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVNKEG0hUkqhBjolBZ1YAoMcoSjlZBJ9e_ggbY4WiSa5dCEPKJtHiyZjcGp3h8QC2LY4w&usqp=CAU",
        "level": 2,
        "rowNo": 6,
        "position": "right",
        "prompt": [
          {
            "type": "text",
            "value": "hi"
          },
          {
            "type": "image",
            "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
          }
        ]
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "rowNo": {
          "type": "integer",
          "format": "int32"
        },
        "position": {
          "type": "string"
        },
        "prompt": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Prompt"
          }
        }
      },
      "required": [
        "name",
        "image",
        "level",
        "rowNo",
        "position",
        "prompt"
      ]
    },
    "createQuestionRequest": {
      "title": "createQuestionRequest",
      "example": {
        "title": "What tempo marking best describes the excerpt you heard?",
        "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVNKEG0hUkqhBjolBZ1YAoMcoSjlZBJ9e_ggbY4WiSa5dCEPKJtHiyZjcGp3h8QC2LY4w&usqp=CAU",
        "audio": "https://www.google.com/",
        "type": "MCQ",
        "options": [
          {
            "option": "moderate",
            "prompt": [
              {
                "type": "text",
                "value": "hi"
              },
              {
                "type": "image",
                "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
              }
            ],
            "correctAnswer": false
          },
          {
            "option": "allegro vivance",
            "prompt": [
              {
                "type": "text",
                "value": "hi"
              },
              {
                "type": "image",
                "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
              }
            ],
            "correctAnswer": false
          },
          {
            "option": "andante",
            "prompt": [
              {
                "type": "text",
                "value": "hi"
              },
              {
                "type": "image",
                "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
              }
            ],
            "correctAnswer": false
          },
          {
            "option": "allegro agetito",
            "prompt": [
              {
                "type": "text",
                "value": "hi"
              },
              {
                "type": "image",
                "value": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQrN7aJrHpL7AqLN7y9V83wSVRZUaPK2h1u8A&usqp=CAU"
              }
            ],
            "correctAnswer": true
          }
        ],
        "musicName": "music",
        "composerName": "composer",
        "topicId": "61fa1aabf65f741f0b6c8a95"
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "audio": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Option"
          }
        },
        "musicName": {
          "type": "string"
        },
        "composerName": {
          "type": "string"
        },
        "topicId": {
          "type": "string"
        }
      },
      "required": [
        "title",
        "image",
        "audio",
        "type",
        "options",
        "musicName",
        "composerName",
        "topicId"
      ]
    },
...
    "createhintrequest": {
      "title": "createhintrequest",
      "example": {
        "hintInfo": {
          "text": "hint info",
          "audio": "https://www.google.com/",
          "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVNKEG0hUkqhBjolBZ1YAoMcoSjlZBJ9e_ggbY4WiSa5dCEPKJtHiyZjcGp3h8QC2LY4w&usqp=CAU"
        },
        "questionInfo": {
          "text": "ques info",
          "audio": "https://www.google.com/",
          "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVNKEG0hUkqhBjolBZ1YAoMcoSjlZBJ9e_ggbY4WiSa5dCEPKJtHiyZjcGp3h8QC2LY4w&usqp=CAU"
        },
        "promptInfo": {
          "text": "prompt info",
          "audio": "https://www.google.com/",
          "image": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTVNKEG0hUkqhBjolBZ1YAoMcoSjlZBJ9e_ggbY4WiSa5dCEPKJtHiyZjcGp3h8QC2LY4w&usqp=CAU"
        },
        "questionId": "61fb797bca3d6a1890f7accc"
      },
      "type": "object",
      "properties": {
        "hintInfo": {
          "$ref": "#/definitions/HintInfo"
        },
        "questionInfo": {
          "$ref": "#/definitions/QuestionInfo"
        },
        "promptInfo": {
          "$ref": "#/definitions/PromptInfo"
        },
        "questionId": {
          "type": "string"
        }
      },
      "required": [
        "hintInfo",
        "questionInfo",
        "promptInfo",
        "questionId"
      ]
    },
...
  },
...
}